requirements.txt
alembic==1.16.5
celery==5.5.3
celery-types==0.23.0
selenium==4.36.0
selenium-stealth==1.0.6
2captcha-python==1.5.1
fastapi[standard]==0.118.3
passlib==1.7.4
pymysql==1.1.2
python_jose==3.5.0
sqlalchemy==2.0.44